move other tests that don't need to run against a remote cluster
