+++ crates/smoketests/tests/smoketests/cli/mod.rs
@@ -1,4 +1,2 @@
-pub mod dev;
-pub mod generate;
 pub mod publish;
 pub mod server;
