+++ crates/smoketests/tests/smoketests/mod.rs
@@ -8,14 +8,9 @@ mod client_connection_errors;
 mod column_defaults;
 mod confirmed_reads;
 mod connect_disconnect_from_cli;
-mod create_project;
-mod csharp_aot_module;
-mod csharp_module;
 mod database_lock;
-mod default_module_clippy;
 mod delete_database;
 mod describe;
-mod detect_wasm_bindgen;
 mod dml;
 mod domains;
 mod fail_initial_publish;
@@ -33,7 +28,6 @@ mod publish_upgrade_prompt;
 mod quickstart;
 mod rls;
 mod schedule_reducer;
-mod servers;
 mod sql;
 mod sql_connect_hook;
 mod templates;
