+++ crates/smoketests/tests/standalone/cli/mod.rs
@@ -1,5 +1,9 @@
 #[path = "../../smoketests/cli/auth.rs"]
 mod auth;
+#[path = "../../smoketests/cli/dev.rs"]
+mod dev;
+#[path = "../../smoketests/cli/generate.rs"]
+mod generate;
 #[path = "../../smoketests/cli/list.rs"]
 mod list;
 #[path = "../cli_server.rs"]
